third_party/blink/web_tests/platform/generic/fast/dom/NodeList/nodelist-namespace-invalidation-expected.txt

This tests that node lists returned by getElementsByTagNameNS are invalidated upon child node changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS appendImageWithId('firstImage'); nodeList = container.getElementsByTagNameNS(namespace, 'img'); nodeList.item(0).id is 'firstImage'
PASS nodeList.length is 1
PASS appendImageWithId('secondImage'); container.removeChild(container.firstChild); nodeList.item(0).id is 'secondImage'
PASS nodeList.length is 1
PASS successfullyParsed is true

TEST COMPLETE
